==> notebooks/02_clean_data.ipynb <==
DF_PATH = "../data/processed/01_preprocessed_df.pkl"
EXPORT_PATH = "../data/processed/02_preprocessed_df.pkl"


# import libraries
import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import seaborn as sns
import pickle


df = pd.read_pickle(DF_PATH)
df.head()

df.info()

-----------

### handiling missing values

df.isnull().sum()

from sklearn.impute import SimpleImputer
imputer = SimpleImputer(strategy='mean')

Separating out the numerical attributes to use the `"median"` strategy (as it cannot be calculated on text attributes like `ocean_proximity`):

numeric_df = df.select_dtypes(include=[np.number])

imputer.fit(numeric_df)

imputer.statistics_

numeric_df.median().values

The imputer has simply computed the median of each attribute and stored the result
in its statistics_ instance variable

Now you can use this “trained” imputer to transform the training set by replacing
missing values by the learned medians

X = imputer.transform(numeric_df)

new_df = pd.DataFrame(X, columns=numeric_df.columns, index=numeric_df.index)
new_df.head()

new_df.shape

new_df.isnull().sum()

-----------

### drop outliers

from sklearn.ensemble import IsolationForest

isolation_forest = IsolationForest(max_samples=100, random_state=42)
pred_outliers = isolation_forest.fit_predict(X)

pred_outliers

new_df = new_df.iloc[pred_outliers == 1]

new_df.shape

As we see that our data drop from 20640 to 17514 rows.

--------

### handiling catergories

new_df.info()

df.info()

cat = df.select_dtypes(include=['object'])

> One issue with `LabelEncoder()` representation is that ML algorithms will assume that two nearby
values are more similar than two distant values. Obviously this is not the case (for
example, categories 0 and 4 are more similar than categories 0 and 1).

To fix this issue, a common solution is to create one binary attribute per category: one attribute
equal to 1 when the category is “<1H OCEAN” (and 0 otherwise), another attribute
equal to 1 when the category is “INLAND” (and 0 otherwise), and so on. This is
called one-hot encoding, because only one attribute will be equal to 1 (hot), while the
others will be 0 (cold).


from sklearn.preprocessing import OneHotEncoder

cat_encoder = OneHotEncoder()
housing_cat_1hot = cat_encoder.fit_transform(cat)

housing_cat_1hot

By default, the `OneHotEncoder` class returns a sparse array, but we can convert it to a dense array if needed by calling the `toarray()` method:

housing_cat_1hot.toarray()

Alternatively, you can set `sparse=False` when creating the `OneHotEncoder`:

cat_encoder = OneHotEncoder(sparse=False)
housing_cat_1hot = cat_encoder.fit_transform(cat)
housing_cat_1hot

housing_cat_1hot.shape

df_output = pd.DataFrame(housing_cat_1hot,
                         columns=cat_encoder.get_feature_names_out(),
                         index=cat.index)

df_output.head()

df_output.shape

data = pd.concat([new_df, df_output], axis=1)
data.head()

data.shape

> we 'll need to drop the NAs from the data which created from the difference in rows in our new_df and df_output

data.isnull().sum()

data1 = data.dropna()

data1.isnull().sum()

data1.shape

data1.info()

----------

### EXPORT

pd.to_pickle(data1, EXPORT_PATH)
