==> notebooks/03_prepare_data_for_ml_models.ipynb <==
DF_PATH = "../data/processed/02_preprocessed_df.pkl"
EXPORT_PATH = "../data/processed/03_X_train_scaled_df.pkl"
EXPORT_PATH1 = "../data/processed/04_X_val_scaled_df.pkl"
EXPORT_PATH2 = "../data/processed/05_y_train_df.pkl"
EXPORT_PATH3 = "../data/processed/06_y_val_df.pkl"
EXPORT_PATH4 = "../data/processed/07_test_df.pkl"
EXPORT_PATH5 = "../data/processed/08_test_scaled_df.pkl"


# import libraries
import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import seaborn as sns
import pickle
from sklearn.pipeline import Pipeline
from sklearn.preprocessing import StandardScaler
from sklearn.preprocessing import MinMaxScaler
from sklearn.model_selection import train_test_split


# load data
df = pd.read_pickle(DF_PATH)
df.head()

--------

### Split data

train, test = train_test_split(df, test_size=0.2, random_state=42)

print(f'--> train: {train.shape}')
print(f'--> test: {test.shape}')

print(train.columns)
print(test.columns)

test_df = test.drop(columns=['median_house_value'])
print(test_df.shape)

##### split train set 

So now le's work with train data and split it into X and y

# removing target variable
X = train.drop('median_house_value', axis=1)
y = train['median_house_value']

X_train, X_val, y_train, y_val = train_test_split(X, y, test_size=0.2, random_state=42)

print(f'--> X_train: {X_train.shape}')
print(f'--> X_val: {X_val.shape}')
print(f'--> y_train: {y_train.shape}')
print(f'--> y_val: {y_val.shape}')

---------

### Feature Scaling

>* Min-max scaling (many people call this normalization) is quite simple: values are shifted and rescaled so that they end up ranging from 0 to 1. We do this by subtracting the min value and dividing by the max minus the min. Scikit-Learn provides a transformer called MinMaxScaler for this. It has a feature_range hyperparameter that lets you change the range if you don’t want 0–1 for some reason.

>* Standardization is quite different: first it subtracts the mean value (so standardized values always have a zero mean), and then it divides by the variance so that the resulting distribution has unit variance.


Unlike min-max scaling, standardization does not bound values to a specific range, which may be a problem for some algorithms (e.g.,
neural networks often expect an input value ranging from 0 to 1). However, standardization is much less affected by outliers.

scaler = StandardScaler()
X_train_scaled = scaler.fit_transform(X_train)
X_val_scaled = scaler.transform(X_val)
test_scaled_df = scaler.transform(test_df)

---------

### EXPORT



# save train data as pickle
pd.to_pickle(X_train_scaled, EXPORT_PATH)
pd.to_pickle(X_val_scaled, EXPORT_PATH1)
pd.to_pickle(y_train, EXPORT_PATH2)
pd.to_pickle(y_val, EXPORT_PATH3)

# save test data as pickle
pd.to_pickle(test_df, EXPORT_PATH4)
pd.to_pickle(test_scaled_df, EXPORT_PATH5)
