==> notebooks/05_fine_tune.ipynb <==
DF_PATH = "../data/processed/03_X_train_scaled_df.pkl"
DF_PATH1 = "../data/processed/04_X_val_scaled_df.pkl"
DF_PATH2 = "../data/processed/05_y_train_df.pkl"
DF_PATH3 = "../data/processed/06_y_val_df.pkl"
DF_PATH4 = "../data/processed/07_test_df.pkl"
DF_PATH5 = "../data/processed/08_test_scaled_df.pkl"


# import libraries
import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import seaborn as sns
import pickle
from sklearn.metrics import mean_squared_error, mean_absolute_error
from sklearn.model_selection import cross_val_score, GridSearchCV
from sklearn.linear_model import LinearRegression
from sklearn.ensemble import RandomForestRegressor


# load data
X_train = pd.read_pickle(DF_PATH)
X_val = pd.read_pickle(DF_PATH1)
y_train = pd.read_pickle(DF_PATH2)
y_val = pd.read_pickle(DF_PATH3)
test = pd.read_pickle(DF_PATH5)


print(X_train.shape, X_val.shape, y_train.shape, y_val.shape, test.shape)

--------

## fine tune

params = {
    "n_estimators": [10, 20, 30],
    "max_features": ["auto", "sqrt", "log2"],
    "max_depth": [2, 3, 4],
    "min_samples_split": [2, 3, 4],
}

forest_reg = RandomForestRegressor()

grid_search = GridSearchCV(forest_reg, params, cv=5,scoring="neg_mean_squared_error")
grid_search.fit(X_train, y_train)

grid_search.best_params_

grid_search.best_estimator_

forest_reg = RandomForestRegressor(**grid_search.best_params_)

# fit the model
forest_reg.fit(X_train, y_train)
# make predictions
y_preds = forest_reg.predict(X_val)

def rmse(y_true, y_preds):
    return np.sqrt(mean_squared_error(y_true, y_preds))


print(f"Validation Score: {forest_reg.score(X_val, y_val)}")
print(f"Validation RMSE: {rmse(y_val, y_preds)}")

> The grid search approach is fine when you are exploring relatively few combinations,like in the previous example, but when the hyperparameter search space is large, it is often preferable to use `RandomizedSearchCV()` instead. This class can be used in much the same way as the GridSearchCV class, but instead of trying out all possible combinations, it evaluates a given number of random combinations by selecting a random value for each hyperparameter at every iteration.

## Evaluate Model on the Test Set

eval_test = forest_reg.predict(test)

# save model
pickle.dump(forest_reg, open("../models/model.pkl", "wb"))

# save predictions
pickle.dump(eval_test, open("../models/predictions.pkl", "wb"))
